# Add lru_cache to some hot spots related to sympy manipulation