--- a/torchinductor/codegen/common.py
+++ b/torchinductor/codegen/common.py
@@ -1,5 +1,6 @@
 import collections
 import contextlib
+import functools
 import itertools
 import logging
 import math
@@ -13,6 +14,7 @@ import sympy
 from sympy.printing.printer import Printer
 
 from .. import metrics
+from ..utils import freeze_inputs
 from ..utils import sympy_dot
 from ..utils import unique
 from ..virtualized import V
@@ -21,6 +23,8 @@ from ..virtualized import ops
 log = logging.getLogger(__name__)
 
 
+@freeze_inputs
+@functools.lru_cache(256)
 def _simplify_loops(index_vars, sizes, index_formulas):
     """
     Try to remove as many axis from loop iterations as possible, by: