--- a/torchinductor/utils.py
+++ b/torchinductor/utils.py
@@ -6,6 +6,8 @@ import numpy as np
 import sympy
 import torch
 from torch.cuda import synchronize
+from torch.fx.immutable_collections import immutable_dict
+from torch.fx.immutable_collections import immutable_list
 
 
 @functools.lru_cache(None)
@@ -75,3 +77,28 @@ def print_performance(fn, args=(), times=10, repeat=10, baseline=1.0):
     took = np.median(timings)
     print(f"{took/baseline:.6f}")
     return took
+
+
+immutable_dict.__hash__ = lambda self: hash(tuple(self.items()))
+immutable_list.__hash__ = lambda self: hash(tuple(self))
+
+
+def freeze_inputs(f):
+    """
+    Useful for wrapping lists in tuples for caching purposes
+    """
+
+    def freeze_value(x):
+        if isinstance(x, list):
+            return immutable_list(x)
+        if isinstance(x, dict):
+            return immutable_dict(x)
+        return x
+
+    @functools.wraps(f)
+    def wrapped(*args):
+        args = [freeze_value(x) for x in args]
+        return f(*args)
+
+    wrapped.cache_info = f.cache_info
+    return wrapped