Reduces compilation time (assuming no autotuning) by about 20-30%.

For ` time  ./benchmarks/torchbench.py -dcuda --inductor --float16 --training --only <model>`

Resnet18: 40s => 30s
densenet121: 2m32s => 2m
hf_Bert: 1m20 => 55s