@Chillee 2022-08-17
I wrapped some more things in `lru_cache` and added a wrapper to `utils.py` to "freeze" inputs.